+++ src/dfwimpl/state_driver.cpp
@@ -276,7 +276,7 @@ void state_driver::prepare_state(int /*_next*/, int /*_current*/) {
 
 }
 
-void state_driver::common_pre_loop_input(dfw::input& input, float _delta) {
+void state_driver::common_pre_loop_input(dfw::input& input, ldtools::tdelta _delta) {
 
 	if(input().is_event_joystick_connected()) {
 		lm::log(log).info()<<"New joystick detected..."<<std::endl;
@@ -286,15 +286,15 @@ void state_driver::common_pre_loop_input(dfw::input& input, float _delta) {
 	message_manager.tick(_delta);
 }
 
-void state_driver::common_loop_input(dfw::input& /*input*/, float /*delta*/) {
+void state_driver::common_loop_input(dfw::input& /*input*/, ldtools::tdelta /*delta*/) {
 
 }
 
-void state_driver::common_pre_loop_step(float /*delta*/) {
+void state_driver::common_pre_loop_step(ldtools::tdelta /*delta*/) {
 
 }
 
-void state_driver::common_loop_step(float /*delta*/) {
+void state_driver::common_loop_step(ldtools::tdelta /*delta*/) {
 
 }